added nav into the footer
