+++ client/src/App.tsx
@@ -9,8 +9,7 @@ import { setContext } from '@apollo/client/link/context';
 import { Outlet } from 'react-router-dom';
 
 import Header from './components/header';
-//import Footer from './components/Footer';
-
+import Footer from './components/footer';
 // Construct our main GraphQL API endpoint
 const httpLink = createHttpLink({
   uri: '/graphql',
@@ -43,10 +42,10 @@ function App() {
         <div className="container">
           <Outlet />
         </div>
-        {/* <Footer /> */}
+        <Footer />
       </div>
     </ApolloProvider>
   );
-}
+};
 
 export default App;
