+++ client/src/components/footer.css
@@ -1,0 +1,11 @@
+footer {
+    margin-top: 80px;
+    
+}
+
+.copyright {
+    position: fixed;
+    bottom: 0;
+    right: 0;
+    background-color: none;
+}
