+++ client/src/main.tsx
@@ -9,6 +9,7 @@ import Login from './components/login.tsx';
 //import SingleThought from './pages/SingleThought';
 //import Profile from './pages/Profile';
 import ErrorPage from './components/error.tsx';
+import AboutArtVine from './pages/AboutArtVine.tsx';
 
 const router = createBrowserRouter([
   {
@@ -34,6 +35,10 @@ const router = createBrowserRouter([
       }, {
         path: '/thoughts/:thoughtId',
         //element: <SingleThought />
+      }, {
+        path: '/AboutArtVine1',
+        element: <AboutArtVine />
+
       }
     ]
   },
@@ -43,3 +48,4 @@ const rootElement = document.getElementById('root');
 if (rootElement) {
   ReactDOM.createRoot(rootElement).render(<RouterProvider router={router} />);
 }
+
